Started to create replicable integration testing environment.
Use of docker container hosting greenmail, mongodb, and the built application.
