+++ .gitignore
@@ -1,3 +1,6 @@
 *.iml
 target/
 .idea/
+.settings/
+.project
+.classpath
