+++ memo-list-magic-core/src/main/java/com/redhat/mlm/IEmailStoreFactory.java
