+++ memo-list-magic-core/src/main/java/com/redhat/mlm/ThreadMetadata.java
