+++ memo-list-magic-core/src/test/java/com/redhat/mlm/MemoListRepoTest.java
