+++ memo-list-magic-core/src/test/java/com/redhat/mlm/ThreadMetadataTest.java
