+++ memo-list-magic-docker/standalone/supervisord.conf
@@ -1,0 +1,19 @@
+[supervisord]
+nodaemon=true
+loglevel=DEBUG
+
+
+[program:greenmail]
+directory=/home/memolister
+command=java -Djava.net.preferIPv4Stack=true -Dgreenmail.setup.test.all -Dgreenmail.hostname=0.0.0.0 -Dgreenmail.auth.disabled -jar greenmail-standalone.jar
+
+[program:mongoDB]
+startsecs=0
+autorestart=false
+directroy=/sbin
+command=mongoDBStartup.sh
+
+
+[program:memolistmagic]
+directory=/home/memolister
+command=java -jar memo-list-magic-standalone.jar ArgOne ArgTwo ArgThree
