+++ pom.xml
@@ -1,71 +1,120 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
-                        http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<?xml version='1.0'?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
+  
+  <name>== MemoListMagic ==</name>
 
   <groupId>com.redhat.mlm</groupId>
-  <artifactId>memo-list-magic</artifactId>
-  <version>1.0-SNAPSHOT</version>
-  <packaging>jar</packaging>
+  <artifactId>memo-list-magic-parent</artifactId>
+  <packaging>pom</packaging>
+  <!-- This version should get inherited to all subprojects. -->
+  <version>1.0.0-SNAPSHOT</version>
 
-  <name>Red Hat Consulting - Memo List Magic</name>
-  <url></url>
+  <modules>
+    <module>memo-list-magic-core</module>
+    <module>memo-list-magic-standalone</module>
+  </modules>
 
-  <properties>
-    <junitVersion>4.12</junitVersion>
-    <javaMailVerion>1.4.7</javaMailVerion>
-    <mockitoVersion>1.10.19</mockitoVersion>
-    <guavaVersion>19.0</guavaVersion>
-    <mongoVersion>3.0.4</mongoVersion>
-  </properties>
+  <build>
+    <defaultGoal>install</defaultGoal>
 
-  <dependencies>
-    <dependency>
-      <groupId>javax.mail</groupId>
-      <artifactId>mail</artifactId>
-      <version>${javaMailVerion}</version>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>${junitVersion}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.mockito</groupId>
-      <artifactId>mockito-all</artifactId>
-      <version>${mockitoVersion}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-		<groupId>com.icegreen</groupId>
-		<artifactId>greenmail</artifactId>
-		<version>1.5.0</version>
-		<scope>test</scope>
-	</dependency>
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-      <version>${guavaVersion}</version>
-    </dependency>
-      <dependency>
-        <groupId>org.mongodb</groupId>
-        <artifactId>mongo-java-driver</artifactId>
-        <version>${mongoVersion}</version>
-      </dependency>
-  </dependencies>
-    <build>
+    <!-- Plugin defaults for all inherited POMs. Declare the plugin version here, too. -->
+    <pluginManagement>
       <plugins>
         <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
+          <!-- http://maven.apache.org/plugins/maven-compile-plugin/plugin-info.html -->
           <artifactId>maven-compiler-plugin</artifactId>
           <version>3.5.1</version>
           <configuration>
             <source>1.8</source>
             <target>1.8</target>
+            <debug>true</debug>
+            <compilerArgs>
+              <arg>-Xlint</arg>
+            </compilerArgs>
+            <showWarnings>true</showWarnings>
+            <showDeprecation>true</showDeprecation>
           </configuration>
         </plugin>
-      </plugins>
-    </build>
+       </plugins>
+     </pluginManagement>
+  </build>
+
+  <properties>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+    <junitVersion>4.12</junitVersion>
+    <javaMailVerion>1.4.7</javaMailVerion>
+    <mockitoVersion>1.10.19</mockitoVersion>
+    <guavaVersion>19.0</guavaVersion>
+    <mongoVersion>3.0.4</mongoVersion>
+    <greenMailVersion>1.5.0</greenMailVersion>
+  </properties>
+
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>com.redhat.mlm</groupId>
+        <artifactId>memo-list-magic</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>com.redhat.mlm</groupId>
+        <artifactId>memo-list-magic-standalone</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+      </dependency>
+    
+      <!-- Greenmail modules -->
+      <dependency>
+        <groupId>com.icegreen</groupId>
+        <artifactId>greenmail</artifactId>
+        <version>${greenMailVersion}</version>
+      </dependency>
+
+      <!-- Compile scope (=default) -->
+      <dependency>
+        <groupId>javax.mail</groupId>
+        <artifactId>mail</artifactId>
+        <version>${javaMailVerion}</version>
+      </dependency>
+      <dependency>
+        <groupId>junit</groupId>
+        <artifactId>junit</artifactId>
+        <version>${junitVersion}</version>
+        <scope>test</scope>
+      </dependency>
+      <dependency>
+        <groupId>org.mockito</groupId>
+        <artifactId>mockito-all</artifactId>
+        <version>${mockitoVersion}</version>
+        <scope>test</scope>
+      </dependency>
+      <dependency>
+  		<groupId>com.icegreen</groupId>
+  		<artifactId>greenmail</artifactId>
+  		<version>${greenMailVersion}</version>
+  		<scope>test</scope>
+  	</dependency>
+      <dependency>
+        <groupId>com.google.guava</groupId>
+        <artifactId>guava</artifactId>
+        <version>${guavaVersion}</version>
+      </dependency>
+        <dependency>
+          <groupId>org.mongodb</groupId>
+          <artifactId>mongo-java-driver</artifactId>
+          <version>${mongoVersion}</version>
+        </dependency>
+    </dependencies>
+  </dependencyManagement>
+
+  <profiles>
+    <profile>
+      <id>docker</id>
+      <modules>
+        <module>memo-list-magic-docker/standalone</module>
+      </modules>
+    </profile>
+  </profiles>
 </project>
